Add files via upload
this is the code for part 3 of project 1.  Basically, it runs options 1, 2, and 3.  It captures the CPU time and creates graphs for each option.  This is done with the tic tac feature in matlab.  The code is pretty basic it just is a bunch of for loops iterating through the various initial conditions.  ie for option 1 there will be more simulations due to there being the k values.  each combination of conditions is then simulated using sim() and the respective simulink model for the option.  Also, the wdot v time graphs are included.  I just did the very last inputted condition set for each option and solver because there would have been about 40 graphs.  and the instructions said " the shaft speeds can be found at each end of each system" so I interpreted that to mean only use the very last condition set. i think I should mostly work.  I had a bit of trouble on the shaft speed v time graphs so idk about those. also, between the different deliverables pdfs given the instructions differ.  like for part 3 deliverables it mentions "Consider constant applied torque of 1 and 100" but for the overall project deliverables it mentions "Consider constant step inputs 1 and 100" so there was difficultly understanding the project.